+++ far/usermenu.cpp
@@ -1098,7 +1098,7 @@ bool UserMenu::EditMenu(std::list<UserMenuItem>& Menu, std::list<UserMenuItem>::
 #else
 		size_t CommandNumber = 0;
 
-		for (const auto i: std::views::iota(size_t{}, DI_EDIT_COUNT))
+		for (const auto i: std::views::iota(size_t{}, static_cast<size_t>(DI_EDIT_COUNT)))
 		{
 			if (!EditDlg[i + EM_EDITLINE_0].strData.empty())
 				CommandNumber = i + 1;
@@ -1106,7 +1106,7 @@ bool UserMenu::EditMenu(std::list<UserMenuItem>& Menu, std::list<UserMenuItem>::
 
 		(*MenuItem)->Commands.clear();
 
-		for (const auto i: std::views::iota(size_t{}, DI_EDIT_COUNT))
+		for (const auto i: std::views::iota(size_t{}, static_cast<size_t>(DI_EDIT_COUNT)))
 		{
 			if (static_cast<size_t>(i) >= CommandNumber)
 				break;
